"{#8} {products.php} {http://pm.5psolutions.net/public/index.php?path_info=projects%2Ftraining-andrei-ghiorghe%2Ftasks%2F8} {This commit updated products.php in order to respect the standards and procedures.}"

+++ products.php
@@ -2,17 +2,12 @@
 require_once 'common.php';
 
 // check if the user is logged in, if he itsn't then I redirect him to login.php
-if (!isset($_SESSION['username']) || $_SESSION['username'] != ADMIN_CREDENTIALS['USERNAME']) {
-    header('Location: login.php');
-    die();
-}
+checkAuthorization();
 
 // if the user wants to logout then
 // I first must delete from session username and after I redirect him to login.php
 if (isset($_POST['logoutButton'])) {
-    unset($_SESSION['username']);
-    header('Location: login.php');
-    die();
+    doLogout();
 }
 
 $connection = databaseConnection();
@@ -30,9 +25,13 @@ if (isset($_POST['deleteItem'])) {
             'DELETE FROM products WHERE id = ?',
             [$_POST['idProductDelete']]
         );
-        // deleting the image from directory
-        unlink($result[0]['image_path']);
+        // deleting the image from directory with checking
+        if (file_exists($result[0]['image_path'])) {
+            unlink($result[0]['image_path']);
+        }
     }
+    header('Location: products.php');
+    die();
 }
 
 $items = query($connection, 'SELECT * FROM products;',[]);
@@ -45,55 +44,54 @@ $items = query($connection, 'SELECT * FROM products;',[]);
     <title> <?= translate('Admin\'s page') ?> </title>
 </head>
 <body>
-
-<table id="contentTable">
-    <tbody>
-    <?php for ($i = 0; $i < count($items); $i++): ?>
-            <tr class="elementOfTable">
-                <td>
-                    <img class="phoneImage" src="<?= $items[$i]['image_path'] ?>">
-                </td>
-                <td>
-                    <?= $items[$i]['title'] ?><br>
-                    <?= $items[$i]['description'] ?><br>
-                    <?= $items[$i]['price'] ?> <?= translate('lei') ?><br>
-                </td>
+    <table id="contentTable">
+        <tbody>
+            <?php for ($i = 0; $i < count($items); $i++): ?>
+                <tr class="elementOfTable">
+                    <td>
+                        <img class="phoneImage" src="<?= $items[$i]['image_path'] ?>">
+                    </td>
+                    <td>
+                        <?= $items[$i]['title'] ?><br>
+                        <?= $items[$i]['description'] ?><br>
+                        <?= $items[$i]['price'] ?> <?= translate('lei') ?><br>
+                    </td>
+                    <td>
+                        <form method="post" action="product.php">
+                            <input type="hidden" name="idProductEdit" value="<?= $items[$i]['id'] ?>">
+                            <button type="submit" class="linkButton" name="editButton">
+                                <?= translate('Edit') ?>
+                            </button>
+                        </form>
+                    </td>
+                    <td>
+                        <form method="post" action="products.php">
+                            <input type="hidden" name="idProductDelete" value="<?= $items[$i]['id'] ?>">
+                            <button type="submit" class="linkButton" name="deleteItem">
+                                <?= translate('Delete') ?>
+                            </button>
+                        </form>
+                    </td>
+                </tr>
+                <br>
+            <?php endfor; ?>
+            <tr>
                 <td>
                     <form method="post" action="product.php">
-                        <input type="hidden" name="idProductEdit" value="<?= $items[$i]['id'] ?>">
-                        <button type="submit" class="linkButton" name="editButton">
-                            <?= translate('Edit') ?>
+                        <button type="submit" name="addButton" class="linkButton">
+                            <?= translate('Add') ?>
                         </button>
                     </form>
                 </td>
                 <td>
                     <form method="post" action="products.php">
-                        <input type="hidden" name="idProductDelete" value="<?= $items[$i]['id'] ?>">
-                        <button type="submit" class="linkButton" name="deleteItem">
-                            <?= translate('Delete') ?>
+                        <button type="submit" class="linkButton" name="logoutButton">
+                            <?= translate('Logout') ?>
                         </button>
                     </form>
                 </td>
             </tr>
-            <br>
-    <?php endfor; ?>
-    <tr>
-        <td>
-            <form method="post" action="product.php">
-                <button type="submit" name="addButton" class="linkButton">
-                    <?= translate('Add') ?>
-                </button>
-            </form>
-        </td>
-        <td>
-            <form method="post" action="products.php">
-                <button type="submit" class="linkButton" name="logoutButton">
-                    <?= translate('Logout') ?>
-                </button>
-            </form>
-        </td>
-    </tr>
-    </tbody>
-</table>
+        </tbody>
+    </table>
 </body>
 </html>
